update to 10.6.4 (4365)

+++ TMessagesProj/src/main/java/org/telegram/ui/TopicCreateFragment.java
@@ -64,7 +64,7 @@ public class TopicCreateFragment extends BaseFragment {
     private final static int EDIT_ID = 2;
     long chatId;
     long selectedEmojiDocumentId;
-    int topicId;
+    long topicId;
 
     TextCheckCell2 checkBoxCell;
     EditTextBoldCursor editTextBoldCursor;
@@ -94,7 +94,7 @@ public class TopicCreateFragment extends BaseFragment {
     @Override
     public boolean onFragmentCreate() {
         chatId = arguments.getLong("chat_id");
-        topicId = arguments.getInt("topic_id", 0);
+        topicId = arguments.getLong("topic_id", 0);
         if (topicId != 0) {
             topicForEdit = getMessagesController().getTopicsController().findTopic(chatId, topicId);
             if (topicForEdit == null) {
